pyproject.toml
[project]
name = "fastapi-chat-clear"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.5",
    "aiokafka>=0.13.0",
    "cohere>=6.1.0",
    "configparser>=7.2.0",
    "dotenv>=0.9.9",
    "fastapi>=0.136.1",
    "fastembed>=0.7.4",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.6",
    "kafka-python>=2.3.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.0",
    "langfuse>=4.5.0",
    "langsmith>=0.7.35",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "pinecone>=8.1.2",
    "pinecone-text>=0.11.0",
    "prometheus-client>=0.25.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.14.0",
    "pymilvus[model]>=2.6.12",
    "python-dotenv>=1.2.2",
    "python-json-logger>=4.1.0",
    "qdrant-client[fastembed]>=1.17.1",
    "rank-bm25>=0.2.2",
    "redis>=7.4.0",
    "sentence-transformers>=5.4.1",
    "sse-starlette>=3.3.4",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "uuid>=1.30",
    "uvicorn[standard]>=0.46.0",
    "weaviate-client>=4.21.0",
]
